+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 For the full commit log, [see here](https://github.com/influxdata/influxdb-ruby/commits/master).
 
+## v0.3.1, released 2016-05-26
+
+- Fixed #130 (again). Integer values are now really written as Integers to InfluxDB.
+
 ## v0.3.0, released 2016-04-24
 
 - Write queries are now checked against 204 No Content responses, in accordance with the official documentation (#128).
@@ -20,7 +24,7 @@ For the full commit log, [see here](https://github.com/influxdata/influxdb-ruby/
 
 - Added `InfluxDB::Client#version`, returning the server version (#117).
 - Fixed escaping issues (#119, #121, #135).
-- Interger values are now written as Integer, not as Float value (#131).
+- Integer values are now written as Integer, not as Float value (#131).
 - Return all result series when querying multiple selects (#134).
 - Made host cycling thread safe (#136).